% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contMatrix.R
\name{contMatrix}
\alias{contMatrix}
\title{Generating a contrast matrix or vector}
\usage{
contMatrix(layout, effect, commRef = FALSE)
}
\arguments{
\item{layout}{experimetnal layout of interest, e.g. "1x3" is for one-factor experiment with three treatments and "2x2" is for two-factor experiments with both the factors with two levels}

\item{effect}{depending on the \code{layout}, \code{effect} could be different, e.g. for one-factor experiment, "all-pair" and "global" can be defined as \code{effect} and for two-factor experiment, effcet could be main effects ("A" or "B"), interaction ("AxB"), and simple effects ("simA" or "simB")}

\item{commRef}{indicates whether the associated design is common reference}
}
\value{
Contrast matrix/vector is returned, whose first two columns/elements correspond to the dye effects
}
\description{
Generates a contrast matrix/vector corresponding to a specific effect
}
\examples{

# For a 1x4 layout, contrast matrix corresponds to all pairwise comparisons
   contMatrix(layout="1x4", effect="all-pair")

# For a 4x3 layout, contrast matrix corresponding to interaction
   contMatrix(layout="4x3", effect="AxB")

}
