% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastEst.R
\name{contrastEst}
\alias{contrastEst}
\title{Estimate and test a contrast matrix/vector}
\usage{
contrastEst(z, dmat, cmat, alternative = "two.sided")
}
\arguments{
\item{z}{vector of responses}

\item{dmat}{design matrix of interest}

\item{cmat}{contrast matrix corresponding to the effect of interest}

\item{alternative}{specification of alternative hypothesis}
}
\value{
Returns a vector of test statistics, estimate of the contrast  matrix/vector, p--value, mean squared error, etc. Each row of the design  matrix corresponds to a response
}
\description{
Estimate and test a contrast matrix/vector
}
\examples{
# design matrix
  dmat <- desMatrix("1x3", design="CL")
  dmat <- rbind(dmat, dmat)

# contraxt vector for comparing first and second factor level
  con <- c(0,0,1,-1,0)

# response
  y <- rnorm(6)
  contrastEst(z=y, dmat=dmat, cmat=con)

}
\references{
Searle SR (1971). Linear Models. Wiley.
}
\seealso{
\code{\link{desMatrix}}, \code{\link{contMatrix}}
}
