% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCriteria.R
\name{eCriteria}
\alias{eCriteria}
\title{Efficiency Criteria}
\usage{
eCriteria(dmat, cmat, cinfo = NULL, m = 0, type = "e",
  cname = NULL)
}
\arguments{
\item{dmat}{design matrix of interest}

\item{cmat}{contrast matrix corresponding to the effect of intererst}

\item{cinfo}{a vector, represents the number of rows corresponding to the
contrast matrices of interest}

\item{m}{number of missing values to be considered for computing
efficiency  criteria}

\item{type}{types of efficiency criteria, e.g. "e" is for E-efficiency, which is the only available option in the current version. Other options such as "d" (D-efficiency) will be added later.}

\item{cname}{names of the contrasts}
}
\value{
A \code{list} of efficiency criteria (\code{eff}) and corresponding
set of missing arrays (\code{m})
}
\description{
Computes efficiency criteria for a given design and the contrast matrix/vector
}
\examples{
# design matrix corresponding to a circular loop design from 3x2 layout
  x <- desMatrix(layout = "3x2", design="CL")

# contrast matrices for the main effects of two factors
   ccA <- contMatrix(layout="3x2", effect="mainA")
   ccB <- contMatrix(layout="3x2", effect="mainB")

# combining the contarst matrices
   cc = rbind(ccA, ccB)

# corresponding value of cinfo
  cin = c(nrow(ccA), nrow(ccB))

# e-optimality without any missing value
  eCriteria(dmat=x, cmat=cc, cinfo=cin, m=0)

# e-efficiency with at most two missing values
  eCriteria(dmat=x, cmat=cc, cinfo=cin, m=2)

}
\references{
Pukelsheim F (1993). Optimal designs of experiments. Wiley.

Landgrebe J, Bretz F, Brunner E (2005). Efficient design and
analysis of two-color factorial microarray design.
Computational Statistics & Data Analysis, 50, 499-517.

Latif AHMM, Bretz F, and Brunner E (2009). Robustness consideration
in selecting efficient two-color microarray designs. Bioinformatics,
25(18), 2355-2361.
}
\seealso{
\code{contMatrix}, \code{desMatrix}, \code{estimable}
}
