% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimable.R
\name{estimable}
\alias{estimable}
\title{Estimability of an effect}
\usage{
estimable(dmat, cmat)
}
\arguments{
\item{dmat}{design matrix of interest}

\item{cmat}{contrast matrix corresponding to the effect of interest}
}
\value{
A logical variable indicating whether the given contrast is
estimable
}
\description{
Check the estimability of an effect
}
\examples{
# design matrix
  dmat <- desMatrix("1x3", design = "CR")

# contraxt vector for comparing first and second factor level
  con <- c(0,0,1,-1,0,0)

# checking estimability
  estimable(dmat, con)


}
\references{
Searle SR (1971). Linear Models. Wiley.
}
\seealso{
\code{contMatrix}, \code{desMatrix}
}
