% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madesign-ga.R
\name{madesign_ga}
\alias{madesign_ga}
\title{Genetic Algorithm for Microarray Designs}
\usage{
madesign_ga(dmat = NULL, layout = NULL, n, cmat, cinfo = NULL,
  Pcross = 0.75, crossType = "one.point", Pmut = NULL,
  selectType = "RSS", verbose = FALSE, wt = NULL, popSize = 50,
  maxGen = 2000, burnOut = 100, convergeNum = 10, nIter = 1,
  m = 0, keep.elite = FALSE, scaling = TRUE, cname = NULL)
}
\arguments{
\item{dmat}{design matrix}

\item{layout}{design layout}

\item{n}{number of available arrays}

\item{cmat}{contrast matrix/vector}

\item{cinfo}{dimension of contrast matrix/vector}

\item{Pcross}{crossover probablity}

\item{crossType}{crossover type, currently available options are: \code{"one.point"}, \code{"two.points"}, and \code{"uniform"}}

\item{Pmut}{mutattion probablity}

\item{selectType}{selection type, currently available options are: sampling proportional to fitness (\code{"SPF"}) and reminder stochastic sampling (\code{"RSS"})}

\item{verbose}{controls the print out during iterations}

\item{wt}{a vector of weights to indicate individual importance of  effects of interest}

\item{popSize}{population size of a generation}

\item{maxGen}{maximum number of generations algorithm will run unless the convergence criteria is met}

\item{burnOut}{number of generations for burning out}

\item{convergeNum}{convergence criteria}

\item{nIter}{number of iterations}

\item{m}{number of missing values to be considered for computing  fitness fuction}

\item{keep.elite}{preserving the best design of each generation}

\item{scaling}{linear scaling on fitness}

\item{cname}{contrast names}
}
\description{
Genetic Algorithms for selecting near-optimal microarray designs for a specific number of arrays. Instead of specifying experimetnal layout, the design matrix corresponding to  the list  of arrays from which near-optimal designs will be searched can also be used.
}
\examples{
# near-optimal designs with 7 arrays for the layout 3x2 and main effects are of interest

# contrast matrix
  A <- contMatrix(layout = "3x2", effect = "mainA")
  B <- contMatrix(layout = "3x2", effect = "mainB")
  cmat <- rbind(A, B)
  cinfo <- c(nrow(A), nrow(B))

\dontrun{
madesign_ga(n=7, cmat=cmat, cinfo=cinfo, layout="3x2")
}

}
\references{
Latif AHM and Brunner E (2016). A genetic algorithm for designing microarray experiments. Computational Statistics, 31(2), 499-424.
}
