% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCriteria.R
\name{rCriteria}
\alias{rCriteria}
\title{robustness criteria}
\usage{
rCriteria(dmat, cmat, cinfo = NULL, type = "e", cname = NULL)
}
\arguments{
\item{dmat}{design matrix of interest}

\item{cmat}{contrast matrix corresponding to the effect of interest}

\item{cinfo}{a vector represents the number of rows corresponding to the
contrast matrices of interest}

\item{type}{specifies the type of the efficiency criteria, either
"a" or "e" or "d"}

\item{cname}{name of the contrasts}
}
\value{
A list of \code{bdn}, \code{avgEff}, and \code{pED}
}
\description{
Computes robustness criteria for a given design and contrast matrix/vector
}
\references{
Latif AHMM, Bretz F, and Brunner E (2009). Robustness consideration
in selecting efficient two-color microarray designs. Bioinformatics,
25(18), 2355-2361.
}
